feat: change github repository dto structures
