+++ apps/server/src/contexts/github/infrastructure/repositories/adapters/github-user.adapter.ts
@@ -1,5 +1,5 @@
 import { Result } from '@/libs/result';
-import { GithubUserDto } from './github-user.dto';
+import { GithubUserDto } from '../dto/github-user.dto';
 
 export function toGithubUserDto(data: unknown): Result<GithubUserDto> {
   try {
