--- apps/server/src/contexts/github/infrastructure/repositories/dto/github-repository.adapter.ts
@@ -1,14 +1,0 @@
-import { Result } from '@/libs/result';
-import { GithubRepositoryDto } from './github-repository.dto';
-
-export function toGithubRepositoryDto(
-  data: unknown,
-): Result<GithubRepositoryDto> {
-  try {
-    const repository = data as { data: GithubRepositoryDto };
-    console.log('repository', repository);
-    return Result.ok(repository.data);
-  } catch (e) {
-    return Result.fail(e);
-  }
-}
