+++ apps/server/src/contexts/github/infrastructure/repositories/inputs/invite-user-to-repo.inputs.dto.ts
@@ -1,4 +1,4 @@
-import { GithubRepositoryPermissionsDto } from './github-permissions.dto';
+import { GithubRepositoryPermissionsDto } from '../dto/github-permissions.dto';
 
 export class InviteUserToRepoInput {
   owner: string;
